+++ packages/shared/src/index.ts
@@ -1,0 +1,5 @@
+export * from "./constants.js";
+export * from "./math/vec3.js";
+export * from "./world/coords.js";
+export * from "./world/chunk.js";
+export * from "./blocks/registry.js";